Move progress update to own module
Scrape frontend should not depend on scrape helpers
Also fixes small bug in new locking
