+++ lmfdb/inventory_app/scrape_frontend.py
@@ -1,7 +1,7 @@
 import scripts.reports.jsonify_db_structure as jdbs
 import inventory_upload_data as iud
 from lmfdb.base import getDBConnection
-from scrape_helpers import update_scrape_progress
+from scrape_progress_update import update_scrape_progress
 import threading
 import bson
